+++ CHANGELOG.md
@@ -1,9 +1,14 @@
+# 0.12.0 
+
+New functionality:
+
+* Docker plugins have by default a new parameter `--data-dir` that can be used to specify where to put the node data
+
 # 0.11.0
 
 New functionality:
 
 * Add node.Remove method for removing nodes (from the cli)
-* Docker plugins have by default a new parameter `--data-dir` that can be used to specify where to put the node data
 
 Bug fixes: